requirements.txt
fastapi~=0.115.5
uvicorn
sqlalchemy~=2.0.36
psycopg2-binary
pydantic~=2.10.2
requests~=2.32.3
python-telegram-bot
alembic
python-dotenv~=1.0.1
python-dateutil~=2.9.0.post0
APScheduler~=3.11.0
bcrypt~=4.2.1
Flask~=3.1.0
config~=0.5.1
aiohttp~=3.11.9
pytest~=8.3.4
passlib~=1.7.4
wtforms~=3.2.1